LayoutTests/js/dfg-mul-big-integer-with-small-integer-and-bitor-expected.txt

Tests what happens when you multiply a big unknown integer with a small known integer and use the result in a bitop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS foo(2147483647) is -65536
PASS successfullyParsed is true

TEST COMPLETE
